+++ tempCodeRunnerFile.rb
@@ -1,5 +1,1 @@
-class Tester
-  include TsvBuddy
-end
-
-t = Tester.new.take_tsv(File.read('data/survey.tsv'))
+File.write('tsv_filename.tsv', t)
