+++ tsv_filename.tsv
@@ -1,0 +1,23 @@
+date	student_id	languages	best_language	app_experience	tech_experience
+9/12/2014 20:20:55	1634	Java, C , HTML , JavaScript , 	Java	Native GUI applications, Web Applications, Mobile Applications, Database Driven Applications	Unix-based OS, Database, Regular Expressions, Web Design
+9/12/2014 20:47:35	8763	C++	C++	Command line applications	Database, Web Design
+9/12/2014 21:14:45	4633	C,verilog,spice,html,javascript,ASP,php,mySQL,android,matlab	php	Command line applications, Web Applications, Database Driven Applications	Unix-based OS, Database, Regular Expressions, Web Design
+9/12/2014 22:03:13	1742	Java,Python,ruby,Javascript,PHP,ActionScript	ruby	Command line applications, Web Applications, Database Driven Applications	Unix-based OS, Version Control, Database, Regular Expressions, Software Testing, Web Design, Cloud Platforms
+9/13/2014 0:12:20	8026	Java,C,Javascript,html	Java	Command line applications, Web Applications, Mobile Applications	Unix-based OS, Database, Web Design
+9/13/2014 8:22:26	6377	C familiar, Python familiar, R and Matlab OK	Python	Command line applications, Math/Statistic Applications	Unix-based OS, Regular Expressions
+9/13/2014 18:27:25	2580	C,html,php(a little)	C	Web Applications	Unix-based OS, Database, Web Design
+9/14/2014 13:37:18	6878	Java, c++	Java	Command line applications, Web Applications, Server Applications, Math/Statistic Applications, Database Driven Applications	Unix-based OS, Database, Software Testing
+9/14/2014 23:06:12	5045	C++,html,css,C#,vb.net,java(basic level)	html,css	Web Applications, Mobile Applications	Database, Regular Expressions, Web Design
+9/14/2014 23:57:58	7663	Java, C, Python, Linux Shell Script	Linux Shell Script	Command line applications, Web Applications, Mobile Applications, Server Applications	Unix-based OS, Regular Expressions
+9/15/2014 17:23:37	6750	Java, HTML, CSS, ASP, .NET.	Java	Native GUI applications, Mobile Applications	Unix-based OS, Database, Web Design
+9/15/2014 17:36:56	1492	C, C++, Javascript, PHP 	C++, PHP	Web Applications	Unix-based OS, Database, Web Design
+9/15/2014 17:55:34	9232	c++	c++	Command line applications	Regular Expressions
+9/15/2014 20:56:16	9717	JAVA , C 	JAVA	Command line applications, Web Applications, Mobile Applications, Database Driven Applications	Unix-based OS, Database, Web Design
+9/16/2014 0:04:12	4938	lots of  C, One semester of JAVA and SQL, project for  PHP HTML	C	Command line applications, Native GUI applications, Web Applications, Server Applications, Database Driven Applications	Unix-based OS, Database, Web Design
+9/17/2014 19:02:03	3632	c#, javascript, html, xaml, c++, cobol, java, vb, cssc	c#	Command line applications, Native GUI applications, Web Applications, Mobile Applications, Database Driven Applications	Unix-based OS, Database, Web Design
+9/18/2014 14:49:50	9513	java, c, linux bash script, python	linux bash script	Command line applications, Web Applications, Server Applications	Unix-based OS
+9/18/2014 20:32:10	6995	java , C , javascript , HTML , CSS	java	Native GUI applications, Web Applications, Mobile Applications, Database Driven Applications	Unix-based OS, Database, Regular Expressions, Web Design
+9/18/2014 22:42:35	7114	Python familiar, C/C++ familiar, R familiar, Matlab rarely use, google script familiar	Python	Command line applications, Native GUI applications, Math/Statistic Applications	Unix-based OS
+9/19/2014 1:42:24	2082	C++, C#, VB.net, HTML+CSS, ASP.net, Java(basic)	HTML+CSS	Web Applications, Mobile Applications	Database, Regular Expressions, Web Design
+9/19/2014 16:25:35	2507	C++, Java, Visual Basic	Java	Command line applications, Native GUI applications, Web Applications, Database Driven Applications	Database, Web Design
+9/20/2014 13:43:24	6551	Java,C++,VB	VB	Command line applications, Native GUI applications, Web Applications, Database Driven Applications	Version Control, Database, Regular Expressions, Software Testing
