+++ yaml_buddy.rb
@@ -3,34 +3,11 @@ require 'yaml'
 
 module YamlBuddy
   def take_yaml(yaml)
-    yml_datas = YAML.safe_load(yaml)
-    @data = ''
-    @data += "#{tsv_title(yml_datas)}\n"
-    yml_datas.each do |datas|
-      tsv_line = ''
-      datas.each do |key, value|
-        tsv_line += key == 'date' ? value.to_s : "\t#{value}"
-      end
-      @data += "#{tsv_line}\n"
-    end
+    @data = YAML.safe_load(yaml)
     @data
-  end
-  
-  def tsv_title(yml_datas)
-    title = ''
-    yml_datas[0].each do |key|
-      title += key[0] == 'date' ? key[0].to_s : "\t#{key[0]}"
-    end
-    title
   end
 
   def to_yaml
-    @data
+    @data.to_yaml
   end
 end
-
-# class Tester
-#   include YamlBuddy
-# end
-
-# t = Tester.new.take_yaml(File.read('data/survey.yml'))
